requirements.txt
torch
torchvision
faiss-cpu
pillow
tqdm
git+https://github.com/openai/CLIP.git